api/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
requests
gunicorn
psycopg2
radar-python